requirements.txt
flet>=0.21.0
fpdf2
openpyxl
